Cleanup and simplification - Remove last OZ dependencies - Simplify deploy and setup scripts - Bridge only first level tokenIds - Fix tests
