+++ contracts/src/investment/CFolioFarm.sol
@@ -8,10 +8,9 @@
 
 pragma solidity 0.7.6;
 
-import '@openzeppelin/contracts/access/Ownable.sol';
-import '@openzeppelin/contracts/math/SafeMath.sol';
-import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import '@openzeppelin/contracts/token/ERC20/SafeERC20.sol';
+import '../../0xerc1155/access/Ownable.sol';
+import '../../0xerc1155/utils/SafeMath.sol';
+import '../../0xerc1155/utils/SafeERC20.sol';
 
 import '../utils/ERC20Recovery.sol';
