+++ contracts/src/investment/Controller.sol
@@ -8,9 +8,9 @@
 
 pragma solidity 0.7.6;
 
-import '@openzeppelin/contracts/access/Ownable.sol';
-import '@openzeppelin/contracts/math/SafeMath.sol';
-import '@openzeppelin/contracts/utils/Context.sol';
+import '../../0xerc1155/access/Ownable.sol';
+import '../../0xerc1155/utils/SafeMath.sol';
+import '../../0xerc1155/utils/Context.sol';
 
 import '../booster/interfaces/IBooster.sol';
 import '../utils/AddressBook.sol';
