+++ contracts/src/investment/UniV2StakeFarm.sol
@@ -8,12 +8,11 @@
 
 pragma solidity 0.7.6;
 
-import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import '@openzeppelin/contracts/token/ERC20/SafeERC20.sol';
-import '@openzeppelin/contracts/math/SafeMath.sol';
-import '@openzeppelin/contracts/access/Ownable.sol';
-import '@openzeppelin/contracts/utils/Context.sol';
-import '@openzeppelin/contracts/utils/ReentrancyGuard.sol';
+import '../../0xerc1155/access/Ownable.sol';
+import '../../0xerc1155/utils/Context.sol';
+import '../../0xerc1155/utils/ReentrancyGuard.sol';
+import '../../0xerc1155/utils/SafeERC20.sol';
+import '../../0xerc1155/utils/SafeMath.sol';
 
 import '../utils/ERC20Recovery.sol';
