+++ contracts/src/polygon/WowsErc1155RootTunnel.sol
@@ -5,6 +5,7 @@ import { IERC1155 } from '../../0xerc1155/interfaces/IERC1155.sol';
 import { ERC1155Holder } from '../../0xerc1155/tokens/ERC1155/ERC1155Holder.sol';
 import { SafeMath } from '../../0xerc1155/utils/SafeMath.sol';
 import { FxBaseRootTunnel } from '../../polygonFx/tunnel/FxBaseRootTunnel.sol';
+import { ISFTEvaluator } from '../cfolio/interfaces/ISFTEvaluator.sol';
 
 import '../token/interfaces/IWOWSCryptofolio.sol';
 import '../token/interfaces/IWOWSERC1155.sol';
@@ -32,6 +33,16 @@ contract WowsERC1155RootTunnel is FxBaseRootTunnel, ERC1155Holder {
 
   IWOWSERC1155 private immutable rootToken_;
   address private immutable childToken_;
+  ISFTEvaluator private immutable sftEvaluator_;
+
+  //////////////////////////////////////////////////////////////////////////////
+  // Modifier
+  //////////////////////////////////////////////////////////////////////////////
+
+  modifier onlyRootToken() {
+    require(msg.sender == address(rootToken_), 'CF: Only from root');
+    _;
+  }
 
   //////////////////////////////////////////////////////////////////////////////
   // Initialization
@@ -41,65 +52,69 @@ contract WowsERC1155RootTunnel is FxBaseRootTunnel, ERC1155Holder {
     address _checkpointManager,
     address _fxRoot,
     address _rootToken,
-    address _childToken
+    address _childToken,
+    ISFTEvaluator _sftEvaluator
   ) FxBaseRootTunnel(_checkpointManager, _fxRoot) {
     require(_rootToken != address(0), 'RT: Invalid root');
     require(_childToken != address(0), 'RT: Invalid child');
 
     rootToken_ = IWOWSERC1155(_rootToken);
     childToken_ = _childToken;
+    sftEvaluator_ = _sftEvaluator;
 
     // MAP_TOKEN, encode(rootToken,uri)
     bytes memory message = abi.encode(MAP_TOKEN, abi.encode(_rootToken));
     _sendMessageToChild(message);
   }
 
-  function deposit(
-    address user,
+  /**
+   * @dev See {IERC1155TokenReceiver-onERC1155Received}
+   */
+  function onERC1155Received(
+    address operator,
+    address from,
     uint256 tokenId,
-    uint256 amount
-  ) public {
-    // Validate ownership
-    require(rootToken_.balanceOf(msg.sender, tokenId) == 1, 'RT: invalid');
-
-    // Transfer from depositor to this contract
-    rootToken_.lockOnChain(tokenId, CHAIN_ID);
-
+    uint256 amount,
+    bytes calldata // data
+  ) public override onlyRootToken returns (bytes4) {
     // Get cfolios
-    bytes memory data = _getCFolio('', tokenId);
+    bytes memory data = _getTokenData('', tokenId);
 
     // DEPOSIT, encode(rootToken, depositor, user, id, amount, extra data)
     bytes memory message = abi.encode(
       DEPOSIT,
-      abi.encode(address(rootToken_), msg.sender, user, tokenId, amount, data)
+      abi.encode(address(rootToken_), operator, from, tokenId, amount, data)
     );
     _sendMessageToChild(message);
+
+    // Call ancestor
+    return super.onERC1155Received(operator, from, tokenId, amount, data);
   }
 
-  function depositBatch(
-    address user,
-    uint256[] memory tokenIds,
-    uint256[] memory amounts
-  ) public {
-    // transfer from depositor to this contract
-    rootToken_.safeBatchTransferFrom(
-      msg.sender, // depositor
-      address(this), // manager contract
-      tokenIds,
-      amounts,
-      ''
-    );
-
+  /**
+   * @dev See {IERC1155TokenReceiver-onERC1155BatchReceived}
+   */
+  function onERC1155BatchReceived(
+    address operator,
+    address from,
+    uint256[] calldata tokenIds,
+    uint256[] calldata amounts,
+    bytes calldata // data
+  ) public override onlyRootToken returns (bytes4) {
     bytes memory data;
     for (uint256 i = 0; i < tokenIds.length; ++i)
-      data = _getCFolio(data, tokenIds[i]);
+      data = _getTokenData(data, tokenIds[i]);
 
     // DEPOSIT_BATCH, encode(rootToken, depositor, user, id, amount, extra data)
     bytes memory message = abi.encode(
       DEPOSIT_BATCH,
-      abi.encode(address(rootToken_), msg.sender, user, tokenIds, amounts, data)
+      abi.encode(address(rootToken_), operator, from, tokenIds, amounts, data)
     );
     _sendMessageToChild(message);
+
+    // Call ancestor
+    return
+      super.onERC1155BatchReceived(operator, from, tokenIds, amounts, data);
   }
 
   //////////////////////////////////////////////////////////////////////////////
@@ -128,18 +143,15 @@ contract WowsERC1155RootTunnel is FxBaseRootTunnel, ERC1155Holder {
       address user,
       uint256 tokenId,
       uint256 amount,
-
-    ) = /*bytes memory data*/
-      abi.decode(
+      bytes memory data
+    ) = abi.decode(
         syncData,
         (address, address, address, uint256, uint256, bytes)
       );
     require(rootToken == address(rootToken_), 'RT: Invalid root');
     require(childToken == childToken_, 'RT: Invalid child');
-    require(amount == 1, 'RT: Invalid amount');
-    require(rootToken_.balanceOf(user, tokenId) == 1, 'RT: Invalid user');
 
-    rootToken_.unlockFromChain(tokenId, CHAIN_ID);
+    rootToken_.safeTransferFrom(address(this), user, tokenId, amount, data);
   }
 
   function _syncBatchWithdraw(bytes memory syncData) internal {
@@ -149,48 +161,33 @@ contract WowsERC1155RootTunnel is FxBaseRootTunnel, ERC1155Holder {
       address user,
       uint256[] memory tokenIds,
       uint256[] memory amounts,
-
-    ) = /*bytes memory data*/
-      abi.decode(
+      bytes memory data
+    ) = abi.decode(
         syncData,
         (address, address, address, uint256[], uint256[], bytes)
       );
     require(rootToken == address(rootToken_), 'RT: Invalid root');
     require(childToken == childToken_, 'RT: Invalid child');
-    require(amounts.length == 0, 'RT: Invalid amounts');
 
-    for (uint256 i = 0; i < tokenIds.length; ++i) {
-      require(rootToken_.balanceOf(user, tokenIds[i]) == 1, 'RT: Invalid user');
-      rootToken_.unlockFromChain(tokenIds[i], CHAIN_ID);
-    }
+    rootToken_.safeBatchTransferFrom(
+      address(this),
+      user,
+      tokenIds,
+      amounts,
+      data
+    );
   }
 
-  function _getCFolio(bytes memory data, uint256 tokenId)
+  function _getTokenData(bytes memory data, uint256 tokenId)
     private
+    view
     returns (bytes memory)
   {
-    rootToken_.lockOnChain(tokenId, CHAIN_ID);
+    (uint256 mintTimestamp, ) = rootToken_.getTokenData(tokenId);
+    uint256 prowess = tokenId.isBaseCard()
+      ? sftEvaluator_.rewardRate(tokenId)
+      : 0;
 
-    if (tokenId.isBaseCard()) {
-      address cfolio = rootToken_.tokenIdToAddress(tokenId);
-      require(cfolio != address(0), 'RT: Invalid cfolio');
-
-      uint256[] memory items = rootToken_.getTokenIds(cfolio);
-      bytes memory result = abi.encodePacked(items.length);
-      // Loop over cfolioItems, remove share, and add them for transfer
-      for (uint256 i = 0; i < items.length; ++i) {
-        // Lock CFI on sidechain
-        rootToken_.lockOnChain(tokenId, CHAIN_ID);
-
-        result = abi.encodePacked(
-          result,
-          items[i],
-          rootToken_.getCFolioItemType(items[i])
-        );
-      }
-      return abi.encodePacked(data, result);
-    } else {
-      return abi.encodePacked(data, rootToken_.getCFolioItemType(tokenId));
-    }
+    return abi.encodePacked(data, mintTimestamp, prowess);
   }
 }
