+++ contracts/src/proxy/UpgradeProxy.sol
@@ -8,11 +8,12 @@
 
 pragma solidity >=0.7.0 <0.8.0;
 
-import '@openzeppelin/contracts/proxy/UpgradeableProxy.sol';
-import '@openzeppelin/contracts/utils/Context.sol';
+import '../../0xerc1155/utils/Context.sol';
 
 import '../utils/AddressBook.sol';
 import '../utils/interfaces/IAddressRegistry.sol';
+
+import './UpgradeableProxy.sol';
 
 contract UpgradeProxy is Context, UpgradeableProxy {
   /**
@@ -81,7 +82,7 @@ contract UpgradeProxy is Context, UpgradeableProxy {
     ifAdmin
   {
     _upgradeTo(newImplementation);
-    Address.functionDelegateCall(newImplementation, data);
+    Address.functionDelegateCall(newImplementation, data, 'Proxy: Call failed');
   }
 
   //////////////////////////////////////////////////////////////////////////////
