+++ package.json
@@ -43,9 +43,9 @@
     "local-node": "hardhat node --network hardhat",
     "local:deploy": "hardhat --network localhost deploy",
     "local:export": "hardhat --network localhost export",
-    "mainnet:deploy": "hardhat --network mainnet deploy --tags Token,TokenSetup,SFT,SFTSetup,TradeFloor,TradeFloorSetup,RewardFarms,RewardFarmsSetup",
-    "mainnet:export": "hardhat --network mainnet export --tags Token,TokenSetup,SFT,SFTSetup,TradeFloor,TradeFloorSetup,RewardFarms,RewardFarmsSetup",
-    "mainnet:verify": "hardhat --network mainnet etherscan-verify",
+    "mainnet:deploy": "hardhat --network mainnet deploy",
+    "mainnet:export": "hardhat --network mainnet export",
+    "mainnet:verify": "hardhat --network mainnet etherscan-verify --solc-input",
     "mumbai:deploy": "hardhat --network mumbai deploy",
     "postinstall": "bash tools/patch-depends.sh",
     "rinkeby:deploy": "hardhat --network rinkeby deploy",
